# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Finance Bot API
  description: |
    Intelligent financial decision simulation platform with multi-agent AI analysis.
    
    This API provides endpoints for:
    - Financial action simulation (save, invest, spend)
    - Multi-action comparison
    - AI-powered agent analysis
    - Sample user data
  version: 1.0.0
  contact:
    name: Finance Bot Team
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server (example)
paths:
...
  /simulate:
    post:
      summary: Simulate a financial action
      description: |
        Simulates a single financial action (save, invest, or spend) and returns
        the projected outcomes including account balances, goal impacts, and budget effects.
      operationId: simulateAction
      tags:
        - Simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateRequest'
            examples:
              save:
                summary: Save $500 to emergency fund
                value:
                  user:
                    $ref: '#/components/examples/sampleUser'
                  action:
                    type: save
                    amount: 500
                    goalId: goal_emergency
              invest:
                summary: Invest $500 in taxable account
                value:
                  user:
                    $ref: '#/components/examples/sampleUser'
                  action:
                    type: invest
                    amount: 500
                    targetAccountId: taxable
                    goalId: goal_house
                    timeHorizon: 5
              spend:
                summary: Spend $100 on groceries
                value:
                  user:
                    $ref: '#/components/examples/sampleUser'
                  action:
                    type: spend
                    amount: 100
                    category: cat_groceries
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /compare:
    post:
      summary: Compare multiple financial actions
      description: |
        Simulates multiple financial actions side-by-side and returns comparison
        results for easy decision-making.
      operationId: compareActions
      tags:
        - Simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
            example:
              user:
                $ref: '#/components/examples/sampleUser'
              options:
                - type: save
                  amount: 500
                  goalId: goal_emergency
                - type: invest
                  amount: 500
                  targetAccountId: taxable
                  goalId: goal_house
                - type: spend
                  amount: 500
                  category: cat_dining
      responses:
        '200':
          description: Array of simulation results for comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimulationResult'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyze:
    post:
      summary: Analyze action with AI agents
      description: |
        Runs a financial simulation and then analyzes it using multi-agent AI system.
        Returns both simulation results and AI-powered recommendations from budgeting,
        investment, guardrail, and validation agents.
      operationId: analyzeAction
      tags:
        - Analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateRequest'
            examples:
              invest_analysis:
                summary: Analyze $500 investment
                value:
                  user:
                    $ref: '#/components/examples/sampleUser'
                  action:
                    type: invest
                    amount: 500
                    targetAccountId: taxable
                    goalId: goal_house
      responses:
        '200':
          description: Combined simulation and analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Analysis failed (e.g., agent error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /demo/scenarios/{id}:
    get:
      summary: Get a specific demo scenario by ID
      description: Returns a single demo scenario with full user profile and suggested actions.
      operationId: getDemoScenarioById
      tags:
        - Demo
      parameters:
        - name: id
          in: path
          required: true
          description: Scenario ID (young_professional, experienced_investor, conservative_saver)
          schema:
            type: string
            enum: [young_professional, experienced_investor, conservative_saver]
      responses:
        '200':
          description: Demo scenario details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoScenario'
        '404':
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /goals/summary:
    post:
      summary: Get detailed goal progress summary
      description: |
        Returns a comprehensive summary of all financial goals including progress,
        status, projections, and suggested actions for each goal.
      operationId: getGoalsSummary
      tags:
        - Goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalsSummaryRequest'
      responses:
        '200':
          description: Goal summary with progress and projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalsSummaryResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Goal summary generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recommend:
    post:
      summary: Get AI-powered financial recommendations
      description: |
        Analyzes the user's financial state, goals, and priorities to generate
        personalized action recommendations. Optionally evaluates recommendations
        with the multi-agent AI system for deeper analysis.
      operationId: getRecommendations
      tags:
        - Analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequest'
            example:
              user:
                $ref: '#/components/examples/sampleUser'
              evaluateWithAgents: false
      responses:
        '200':
          description: Financial health analysis and recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Recommendation generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    SimulateRequest:
      type: object
      required:
        - user
        - action
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        action:
          $ref: '#/components/schemas/FinancialAction'
    CompareRequest:
      type: object
      required:
        - user
        - options
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        options:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FinancialAction'
    SimulationResult:
      type: object
      required:
        - action
        - scenarioIfDo
        - scenarioIfDont
        - confidence
        - reasoning
        - validationResult
      properties:
        action:
          $ref: '#/components/schemas/FinancialAction'
        scenarioIfDo:
          $ref: '#/components/schemas/Scenario'
        scenarioIfDont:
          $ref: '#/components/schemas/Scenario'
        confidence:
          type: string
          enum: [high, medium, low]
        reasoning:
          type: string
        validationResult:
          $ref: '#/components/schemas/ValidationResult'
...
    AnalysisResponse:
      type: object
      required:
        - simulation
        - analysis
        - metadata
      properties:
        simulation:
          $ref: '#/components/schemas/SimulationResult'
        analysis:
          $ref: '#/components/schemas/AgentAnalysisResult'
        metadata:
          $ref: '#/components/schemas/AnalysisMetadata'
...
    DemoScenario:
      type: object
      required:
        - id
        - name
        - persona
        - description
        - highlights
        - user
        - suggestedActions
      properties:
        id:
          type: string
          example: young_professional
        name:
          type: string
          example: Sarah Chen - Young Professional
        persona:
          type: string
          example: 28yo Software Engineer
        description:
          type: string
        highlights:
          type: array
          items:
            type: string
          example:
            - Building emergency fund (53% complete)
            - Saving for house down payment
        user:
          $ref: '#/components/schemas/UserProfile'
        suggestedActions:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedAction'
...
    GoalsSummaryRequest:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
    GoalsSummaryResponse:
      type: object
      required:
        - summary
        - goals
        - metadata
      properties:
        summary:
          $ref: '#/components/schemas/GoalsAggregate'
        goals:
          type: array
          items:
            $ref: '#/components/schemas/GoalDetail'
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
    RecommendRequest:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        evaluateWithAgents:
          type: boolean
          description: Whether to evaluate recommendations with AI agents (slower but more thorough)
          default: false
    RecommendResponse:
      type: object
      required:
        - financialHealth
        - recommendations
        - metadata
      properties:
        financialHealth:
          $ref: '#/components/schemas/FinancialHealth'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        metadata:
          $ref: '#/components/schemas/RecommendMetadata'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: object
          description: Validation error details (if applicable)
        message:
          type: string
          description: Error message (for 500 errors)
  examples:
    sampleUser:
      value:
        id: user_sarah_demo
        name: Sarah
        monthlyIncome: 5000
        accounts:
          checking: 3000
          savings: 8000
          investments:
            taxable: 5000
            rothIRA: 0
            traditional401k: 0
        fixedExpenses:
          - id: exp_rent
            name: Rent
            amount: 1500
            frequency: monthly
            dueDay: 1
        spendingCategories:
          - id: cat_groceries
            name: Groceries
            monthlyBudget: 400
            currentSpent: 0
            transactions: []
        goals:
          - id: goal_emergency
            name: Emergency Fund
            targetAmount: 15000
            currentAmount: 8000
            deadline: '2027-01-31T00:00:00Z'
            priority: 1
            timeHorizon: short
            linkedAccountIds: [savings]
        preferences:
          riskTolerance: moderate
          liquidityPreference: medium
          guardrails:
            - id: guard_checking_min
              rule: Never let checking drop below $1,000
              type: min_balance
              accountId: checking
              threshold: 1000
        createdAt: '2026-01-01T00:00:00Z'
        updatedAt: '2026-01-31T00:00:00Z'
...